requirements.txt
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygan==1.0.8
pylab-sdk==1.1.2
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
playsound==1.2.2
pyautogui==0.9.50
tkmacosx==0.1.3
pathlib==1.0.1